+++ webpack.config.js
@@ -2,12 +2,13 @@
 'use strict';
 
 const path = require('path');
+const webpack = require('webpack');
 
 // @ts-check
 /** @typedef {import('webpack').Configuration} WebpackConfig **/
 
 /** @type WebpackConfig */
-const extensionConfig = {
+const nodeConfig = {
     target: 'node',
     mode: 'none',
 
@@ -22,6 +23,7 @@ const extensionConfig = {
         vscode: 'commonjs vscode'
     },
     resolve: {
+        mainFields: ['module', 'main'],
         extensions: ['.ts', '.js']
     },
     module: {
@@ -42,4 +44,55 @@ const extensionConfig = {
         level: 'log'
     }
 };
-module.exports = [extensionConfig];
+
+/** @type WebpackConfig */
+const webConfig = {
+    target: 'webworker',
+    mode: 'none',
+
+    entry: './src/extension.ts',
+    output: {
+        path: path.resolve(__dirname, 'dist'),
+        filename: 'extension.js',
+        libraryTarget: 'commonjs2'
+    },
+    externals: {
+        // Modules added here also need to be added in the .vscodeignore file.
+        vscode: 'commonjs vscode'
+    },
+    resolve: {
+        mainFields: ['browser', 'module', 'main'],
+        extensions: ['.ts', '.js'],
+        alias: {},
+        fallback: {
+            util: require.resolve('util')
+        }
+    },
+    module: {
+        rules: [
+            {
+                test: /\.ts$/,
+                exclude: /node_modules/,
+                use: [
+                    {
+                        loader: 'ts-loader'
+                    }
+                ]
+            }
+        ]
+    },
+    plugins: [
+        new webpack.ProvidePlugin({
+            Buffer: ['buffer', 'Buffer']
+        }),
+        new webpack.ProvidePlugin({
+            process: 'process/browser'
+        })
+    ],
+    devtool: 'nosources-source-map',
+    infrastructureLogging: {
+        level: 'log'
+    }
+};
+
+module.exports = [nodeConfig, webConfig];